iconmanager: Uses external icon repository
Code was modified to use an externally configured icon store. The primary icon
information source is the icon files in svg format (and saved in the svg
directory).

Metadata is saved in a separate file using the sha1 hash of the image file as
the key. It is not critical that a metadata file exists, but once it does exist
it should be preserved as it will contain information that cannot be recovered
if lost.